% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{check_stan_warnings}
\alias{check_stan_warnings}
\title{Check Stan warnings}
\usage{
check_stan_warnings(stan_fit, verbose = TRUE)
}
\arguments{
\item{stan_fit}{Fitted Stan model}

\item{verbose}{Print warnings?}
}
\description{
Checks for convergence issues and prints warnings.
}
\keyword{internal}
