% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{get_default_parameters}
\alias{get_default_parameters}
\title{Get default parameters to extract}
\usage{
get_default_parameters()
}
\description{
Returns standard parameters for EHT model.
}
\keyword{internal}
